requirements.txt
aiocache==0.11.1
aiodns==2.0.0
aiofiles==0.5.0
aiohttp==3.6.2
aiomysql==0.0.20
aioredis==1.3.1
aiosqlite==0.15.0
aiounittest==1.4.0
APScheduler==3.6.3
async-timeout==3.0.1
asyncinit==0.2.4
asyncpg==0.21.0
attrs==20.2.0
backcall==0.2.0
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
contextvars==2.4
cryptography==3.1.1
decorator==4.4.2
dialogflow==1.1.0
dialogflow-fulfillment==0.4.0
enum34==1.1.10
Flask==1.1.2
google-api-core==1.22.2
google-auth==1.21.1
googleapis-common-protos==1.52.0
grpcio==1.32.0
h11==0.9.0
hiredis==1.1.0
httpcore==0.10.2
httpx==0.14.3
idna==2.10
immutables==0.14
iniconfig==1.0.1
ipython==7.18.1
ipython-genutils==0.2.0
iso8601==0.1.13
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
loguru==0.5.2
MarkupSafe==1.1.1
more-itertools==8.5.0
multidict==4.7.6
packaging==20.4
parso==0.7.1
peewee==3.13.3
peewee-async==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.7
protobuf==3.13.0
psycopg2==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
pydantic==1.6.1
Pygments==2.7.0
PyMySQL==0.9.2
pyparsing==2.4.7
PyPika==0.39.1
pytest==6.0.2
pytest-asyncio==0.14.0
python-dotenv==0.14.0
pytz==2020.1
rapidfuzz==0.11.2
requests==2.24.0
rfc3986==1.4.0
rsa==4.6
six==1.15.0
sniffio==1.1.0
SQLAlchemy==1.3.19
toml==0.10.1
tortoise-orm==0.16.15
traitlets==5.0.4
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==3.2.0
urllib3==1.25.10
uvloop==0.14.0
vbml==0.5.93
vkbottle==2.7.12
watchgod==0.6
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
